+++ pydp/__init__.py
@@ -1,3 +1,5 @@
 from ._pydp import *
+from pydp import algorithms
+
 
 __version__ = "0.1.4"
