+++ pydp/algorithms/laplacian/__init__.py
@@ -1,0 +1,3 @@
+from .count import Count
+
+__all__ = ["Count"]
